# packages/contracts/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DamdAImin API",
    "description": "Versioned application API for authenticated DamdAImin journeys.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v2/analysis-uploads": {
      "post": {
        "summary": "Create a private Source Audio upload operation",
        "tags": [
          "Analyses"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "language",
                  "contractVersion"
                ],
                "properties": {
                  "language": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "taglish"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "english"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "tagalog"
                        ]
                      }
                    ]
                  },
                  "contractVersion": {
                    "minLength": 1,
                    "maxLength": 64,
                    "type": "string"
                  },
                  "retainSourceAudio": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "uploadId",
                    "uploadUrl",
                    "uploadMethod",
                    "uploadHeaders",
                    "expiresAt"
                  ],
                  "properties": {
                    "uploadId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "uploadUrl": {
                      "format": "uri",
                      "type": "string"
                    },
                    "uploadMethod": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "PUT"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "POST"
                          ]
                        }
                      ]
                    },
                    "uploadHeaders": {
                      "type": "object",
                      "required": [
                        "content-type"
                      ],
                      "properties": {
                        "content-type": {
                          "type": "string",
                          "enum": [
                            "audio/wav"
                          ]
                        }
                      }
                    },
                    "expiresAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "validation_error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized"
                      ]
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "service_unavailable"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/analysis-uploads/{uploadId}/transcription-preview": {
      "post": {
        "summary": "Generate an editable transcript preview for uploaded Source Audio",
        "tags": [
          "Analyses"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "uploadId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "transcript"
                  ],
                  "properties": {
                    "transcript": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error",
                    "message"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "validation_error"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "unauthorized"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "not_found"
                      ]
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "service_unavailable"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ]
}